chore: extract config funcs to separate file
