+++ cd/main_test.go
@@ -4,6 +4,8 @@ import (
 	"testing"
 )
 
+const TEST_PROJECT = "cd-test"
+
 func TestRunVersion(t *testing.T) {
 	if run("cd", "--version") != 0 {
 		t.Error("run --version flag should exit with 0")
@@ -17,7 +19,7 @@ func TestRunUsageError(t *testing.T) {
 }
 
 func TestRunHelp(t *testing.T) {
-	t.Setenv("PROJECT", t.Name())
+	t.Setenv("PROJECT", TEST_PROJECT)
 	if run("cd", "help") != 0 {
 		t.Error("run help flag should exit with 0")
 	}
